refactor: chip types
